pyproject.toml
[project]
name = "ai-task-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "gitpython>=3.1.45",
    "google-generativeai>=0.8.6",
    "langchain-google-genai>=4.1.1",
    "langgraph>=1.0.5",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.9",
    "ty>=0.0.3",
]
